% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_points.R
\name{sample_points}
\alias{sample_points}
\title{Sample points}
\usage{
sample_points(x, filter = "-keep_class 2 -keep_random_fraction 0.01")
}
\arguments{
\item{x}{Virtual point cloud, LAScatalog, LAS, or path to files}

\item{filter}{Character. Filters passed to lasR::reader_las}
}
\value{
Dataframe containing XYZ coordinates of sampled points
}
\description{
Sample points
}
\examples{
f <- system.file("extdata", "Topography.las", package = "lasR")
df <- sample_points(f)
head(df)
}
