% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifidar-package.R
\docType{package}
\name{verifidar-package}
\alias{verifidar}
\alias{verifidar-package}
\title{verifidar: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://samherniman.github.io/verifidar/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
